Fix avatar upload forever

+++ pinterest-server/app/Http/Controllers/Api/UserController.php
@@ -181,22 +181,27 @@ class UserController extends BaseController
     }
 
     public function uploadUserAvatar(Request $request){
-        $user = User::where('name', $request['name'])->first();
-        if($user->avatar_url !== "/img/avatar.jpeg"){
-            unlink(substr($user->avatar_url, 1));
+        if($request["imgBase64"] === "" || $request["imgBase64"] === null){
+            return response()->json(["message" => "no file"], 404);
+        }
+        else{
+            $user = User::where('name', $request['name'])->first();
+            if($user->avatar_url !== "/img/avatar.jpeg"){
+                unlink(substr($user->avatar_url, 1));
+            }
+
+            $output_file = "img/". $request["name"] ."Avatar". ((new DateTime())->format('Y-m-d--H:i:s')).".jpeg";
+            $ifp = fopen( $output_file, 'wb' );
+            $data = explode( ',', $request["imgBase64"] );
+            fwrite( $ifp, base64_decode( $data[ 1 ] ) );
+            fclose( $ifp );
+
+            $user = User::where('name', $request['name'])->first();
+            $user->avatar_url = "/" . $output_file;
+            $user->save();
+            return response()->json($output_file);
         }
 
-        $output_file = "img/". $request["name"] ."Avatar". ((new DateTime())->format('Y-m-d--H:i:s')).".jpeg";
-        $ifp = fopen( $output_file, 'wb' );
-        $data = explode( ',', $request["imgBase64"] );
-        fwrite( $ifp, base64_decode( $data[ 1 ] ) );
-        fclose( $ifp );
-
-        $user = User::where('name', $request['name'])->first();
-        $user->avatar_url = "/" . $output_file;
-        $user->save();
-
-        return response()->json($output_file);
     }
 
     public function getUserAvatar(Request $request){
